+++ packages/components/organisms/f-header/test/visual/f-header.visual.mobile.spec.js
@@ -1,7 +1,5 @@
 import forEach from 'mocha-each';
 
-import { userInfo } from '../../test-utils/helpers/objects';
-
 const Header = require('../../test-utils/component-objects/f-header.component');
 
 let header;
@@ -14,48 +12,53 @@ describe('Shared - f-header component tests', () => {
         ['en-IE', true], ['en-IE', false],
         ['it-IT', true], ['it-IT', false],
         ['es-ES', true], ['es-ES', false]
-    ])
-        .it('should display component', (tenant, isLoggedIn) => {
-            // Arrange
-            header = new Header()
-                .withQuery('&knob-Locale', tenant)
-
-                // Both props below should only show for UK
-                .withQuery('&knob-Show offers link', 'true')
-                .withQuery('&knob-Show delivery enquiry', 'true')
-
-                .withQuery('&knob-Is logged in?', isLoggedIn);
-
-            if (isLoggedIn) {
-                header.withQuery('&knob-User info (if logged in)', JSON.stringify(userInfo));
-            }
-
-            // Act
-            header.load();
-            header.openMobileNavigationBar();
-
-            // Assert
-            browser.percyScreenshot(`f-header - Base state - isLoggedIn: ${isLoggedIn} - ${tenant}`, 'mobile');
-        });
-
-    forEach(['white', 'highlight', 'transparent'])
-        .it('should display the "%s" header theme', theme => {
-            // Arrange
-            header = new Header()
-                .withQuery('&knob-Locale', 'en-GB')
-                .withQuery('&knob-Header theme', theme);
-
-            // Act
-            header.load();
-
-            // Assert
-            browser.percyScreenshot(`f-header - Theme colours - ${theme}`, 'mobile');
-        });
-
-    it('should display all avalible countries', () => {
+    ]).it('should display component', (tenant, isLoggedIn) => {
         // Arrange
-        header = new Header()
-            .withQuery('&knob-Locale', 'en-GB');
+        const controls = [
+            `locale:${tenant}`,
+            'showOffersLink:true', // Should show for AU, IE, NZ and UK only
+            'showDeliveryEnquiry:true', // Should show for AU, IE, NZ and UK only
+            ...(isLoggedIn ? [] : ['userInfoProp:!undefined'])
+        ].join(';');
+
+        header = new Header();
+        header.path += `&args=${controls}`;
+
+        // Act
+        header.load();
+        header.openMobileNavigationBar();
+
+        // Assert
+        browser.percyScreenshot(`f-header - Base state - isLoggedIn: ${isLoggedIn} - ${tenant}`, 'mobile');
+    });
+
+    forEach([
+        'white',
+        'highlight',
+        'transparent'
+    ]).it('should display the "%s" header theme', theme => {
+        // Arrange
+        const controls = [
+            'locale:en-GB',
+            `headerBackgroundTheme:${theme}`
+        ].join(';');
+
+        header = new Header();
+        header.path += `&args=${controls}`;
+
+        // Act
+        header.load();
+
+        // Assert
+        browser.percyScreenshot(`f-header - Theme colours - ${theme}`, 'mobile');
+    });
+
+    it('should display all available countries', () => {
+        // Arrange
+        const controls = 'locale:en-GB';
+
+        header = new Header();
+        header.path += `&args=${controls}`;
 
         // Act
         header.load();
@@ -66,28 +69,42 @@ describe('Shared - f-header component tests', () => {
         browser.percyScreenshot('f-header - Country list', 'mobile');
     });
 
-    forEach(['Show login/user info link', 'Show help link', 'Show country selector'])
-        .it('should not display "%s" ', knobName => {
-            // Arrange
-            header = new Header()
-                .withQuery('&knob-Locale', 'en-GB')
-                .withQuery(`&knob-${knobName}`, 'false');
-
-            // Act
-            header.load();
-            header.openMobileNavigationBar();
-
-            // Assert
-            browser.percyScreenshot(`f-header - ${knobName} - False`, 'mobile');
-        });
-
-    it('should display any custom links', () => {
+    forEach([
+        'showLoginInfo',
+        'showHelpLink',
+        'showCountrySelector'
+    ]).it('should not display "%s" ', knobName => {
         // Arrange
-        header = new Header()
-            .withQuery('&knob-Locale', 'en-GB')
-            .withQuery('&knob-Show offers link', false)
-            .withQuery('&knob-Show delivery enquiry', false)
-            .withQuery('&knob-Show custom nav links?', true);
+        const controls = [
+            'locale:en-GB',
+            `${knobName}:false`
+        ].join(';');
+
+        header = new Header();
+        header.path += `&args=${controls}`;
+
+        // Act
+        header.load();
+        header.openMobileNavigationBar();
+
+        // Assert
+        browser.percyScreenshot(`f-header - ${knobName} - False`, 'mobile');
+    });
+
+    // Not currently possible to set complex values (i.e., arrays) for controls via query strings.
+    // https://storybook.js.org/docs/vue/essentials/controls#dealing-with-complex-values
+    // https://github.com/storybookjs/storybook/issues/14420
+    it.skip('should display any custom links', () => {
+        // Arrange
+        const controls = [
+            'locale:en-GB',
+            'showOffersLink:false',
+            'showDeliveryEnquiry:false'
+            // Set custom links here
+        ].join(';');
+
+        header = new Header();
+        header.path += `&args=${controls}`;
 
         // Act
         header.load();
@@ -97,17 +114,23 @@ describe('Shared - f-header component tests', () => {
         browser.percyScreenshot('f-header - with custom nav links', 'mobile');
     });
 
-    it('should be able to show only custom links', () => {
+    // Not currently possible to set complex values (i.e., arrays) for controls via query strings.
+    // https://storybook.js.org/docs/vue/essentials/controls#dealing-with-complex-values
+    // https://github.com/storybookjs/storybook/issues/14420
+    it.skip('should be able to show only custom links', () => {
         // Arrange
-        header = new Header()
-            .withQuery('&knob-Locale', 'en-GB')
-            .withQuery('&knob-Show help link', false)
-            .withQuery('&knob-Show offers link', false)
-            .withQuery('&knob-Show country selector', false)
-            .withQuery('&knob-Show delivery enquiry', false)
-            .withQuery('&knob-Show login/user info link', false)
+        const controls = [
+            'locale:en-GB',
+            'showHelpLink:false',
+            'showLoginInfo:false',
+            'showOffersLink:false',
+            'showCountrySelector:false',
+            'showDeliveryEnquiry:false'
+            // Set custom links here
+        ].join(';');
 
-            .withQuery('&knob-Show custom nav links?', true);
+        header = new Header();
+        header.path += `&args=${controls}`;
 
         // Act
         header.load();